reusable component for info footer  and inserted in respective places
